requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
duckduckgo-search
slowapi
python-jose[cryptography]
python-multipart
jinja2
